add: break main file into functional pieces
